requirements.txt
django-bootstrap3
pillow
django-multiselectfield
django-misaka
django-slugify-processor
